Prepare `filesystem_caches` directories in the server entrypoint
The entrypoint pre-creates the directories the server needs, but reads only
`storage_configuration.disks.*`. A `filesystem_caches` entry is not a disk, so
the server has to create that directory itself in `FileCache::initialize`, which
fails when the parent is a directory it does not own:

    Code: 36. DB::Exception: Failed to retrieve filesystem information for cache
    path /mnt/clickhouse-cache/stateless_worker_cache. Error: filesystem error:
    in create_directories: Permission denied. (BAD_ARGUMENTS)

Reproduced on the released image with an absolute cache path under a
`root:root 0755` mount. So read `filesystem_caches.*.path` beside the disk paths
and prepare them in the same loop.

Absolute entries only. `FileCacheSettings::loadFromConfig` uses an absolute
`path` verbatim but prefixes a relative one, and that prefix is the server's to
choose, so preparing a relative value as written would create a directory the
server never opens.

Three docker-library cases, discovered automatically by `config.sh`. The
absolute case fails on the stock image and passes here. The relative case fails
if the absolute-only filter is dropped. The third configures
`filesystem_caches_path` with a relative entry and asserts only that the server
starts and owns whichever directory it reports, with no claim about the location:
that keeps it green if the server's initialization order is ever corrected
together with the matching preparation, and red if such a change leaves the new
location unprepared.

The docker-library runner moves out of `ci/jobs/docker_server.py` into
`ci/jobs/scripts/docker_server/docker_library.py` so another job can reuse it
without importing a job entry point, which would pull in `ci.defs.job_configs`
and the rest of that script's module scope. It resolves its paths from `__file__`
rather than the working directory, and takes the image list as an optional
argument, deriving it as before when omitted, so this job is unchanged.

Also quotes `basename "$USERS_XML"`, which shellcheck flags and which the WizOS
fork of this file had already fixed.
