+++ ci/jobs/scripts/docker_server/tests/clickhouse-filesystem-cache-caches-path/cache.xml
@@ -1,0 +1,9 @@
+<clickhouse>
+    <filesystem_caches_path>/mnt/clickhouse-cache/</filesystem_caches_path>
+    <filesystem_caches>
+        <docker_relative_cache>
+            <max_size>10000000</max_size>
+            <path>docker_relative_cache</path>
+        </docker_relative_cache>
+    </filesystem_caches>
+</clickhouse>
